Rearrange loop, fix subsetting files
For loop was closed too early. The result was only the last year was
cleaned and reformatted. This fixes that issue.

Correct some issues with subsetting based on Country or
Agroclimatology. If a country is selected, the whole year .gz file is
downloaded and unzipped and then filtered, rather than individual
files. This prevents issues with too many server connections, however
it is slower, but is more likely work properly.

Some other house cleaning and clarifications.