Fix bugs in get_country
get_country failed when ISO code was not first three letters of country
name.

This update allows for the use of both 2 and 3 letter ISO codes
directly from the user and also finds the proper country name if the
ISO 3 letter code is not the first three letters of the country name.