Bug fixes and fix for RH > 100%
This commit fixes the issue where RH > 100%. The error was caused by
using readr::read_table, which was not able to properly read the input
files.

Using readr::read_fwf() has rectified the situation.

Other fixes include spelling error corrections and
setting MIN/MAX to NA where one or the other is NA.

New columns included in the output include the MIN/MAX flags for how
the value was determined, previously this was dropped.