Add shapefile option for final file output
This commit adds the option for the user to save output as an ESRI
format shapefile in addition to a CSV.  This change forces and update
to stations.rda and fetch_isd-history.R as well for the previously
named STATION.NAME field, now STN.NAME.

Due to this addition, field names must be <10 characters. All field
names >10 chars have been changes/shortened from this release on.

Remove tidyr from imports as it's no longer used with switch to
readr::read_fwf. Column parsing is done on import, negating the need
for any tidyr functions.