Major bug fixes for precipitation values
Fix conversion bug which causes precipitation and snow depth to be too
large.

Also fix bug where NAs in original GSOD data were not recognised. 99.99
is also used as a "missing value" in original data. This was not
included in the string of values on line 488 of `get_GSOD.R`