Bugfixes and enhancements
If a station or several stations are queried, the function now checks
the years that NCDC reports stations having data and gives an error
message if outside this range.

In some cases, this is inaccurate or files simply do not exist. If this
is the case, the function simply moves on and silently fails when there
are no files to ftp from NCDC. That is, if a file is requested that
does not exist, the function returns a message.

`nearest_stations()` is now functional

Update documentation for all functions and files

Added RCUrl to dependencies

Fixed typos in fetch_isd-history.Rmd

Update vignette with more examples, especially using
`nearest_stations()`