Provide only SRTM 90m elevation data
Rather than providing our own version of the isd-history.csv file,
instead we will provide a data frame that includes only a STNID to join
the CGIAR_SRTM elevation data column with the idd-history.csv file,
which will be fetched when running the `get_GSOD()` function.

This means that the latest data for the stations should be available
including the dates available, new stations coming online, etc. without
the need for frequent updates to the R package. New stations come
online much less frequently, which mean that the elevation values do
not need to be updated as often.

This negates the need for the "refresh" option in `get_GSOD()` as well.