Check to see if max_missing specified/set up parallel at start
If specified, run check on files. This adds to the time it takes to run.

Set up parallel options before processing any files. This allows us to
use it any nested functions within get_GSOD, not just .process_files.
We can call it in the llply call to check file lengths for max_missing.