Update paper.md
Edit for changes in data included in the package. The climate data was removed to the `GSODRdata` package available from GitHub, https://github.com/adamhsparks/GSODRdata due the installed package size with the climate data included being too large for CRAN policies.

The data are referenced in the documentation but not necessary for the `GSODR` package to work.