Write test for and fix bug where options("timeout") was not reset
`options(“timeout”)` is set to 300 to interact with NCDC server.

It should reset to 60 but it was not doing so.

New test written to check for this as well as fixing the bug

Minor fixes in `get_GSOD.R` to remove `cache_dir` object

Remove check for stations list, impractical to implement at this point