Correct local image issues
“It was recently pointed out to us that some README.html files
(generated
from the corresponding README.md ones) on the CRAN package web pages are
incomplete, missing 'local' images not available from the web page and
in most cases actually not even shipped with the package.  This clearly
should be changed, so we will move to using '--self-contained' for the
pandoc conversion to ensure that the README.html files are "complete".

Of course, this implies that all 'local' images used in README.md are
needed in the package sources.

If the images are also used for vignettes or Rd files, you can put them
in the 'vignettes' or 'man/figures' directories.  Otherwise, please put
them in the top-level 'tools' directory, or a subdirectory of it.”

The README-example_3.2.-1.png has been removed and README.md now uses a
copy of the same figure from `vignettes/figure/example_1.2-1.png` to
correct the issue.