Correct missing image in CRAN readme.html file
“It was recently pointed out to us that some README.html files
(generated
from the corresponding README.md ones) on the CRAN package web pages are
incomplete, missing 'local' images not available from the web page and
in most cases actually not even shipped with the package.  This clearly
should be changed, so we will move to using '--self-contained' for the
pandoc conversion to ensure that the README.html files are "complete".

Of course, this implies that all 'local' images used in README.md are
needed in the package sources.

If the images are also used for vignettes or Rd files, you can put them
in the 'vignettes' or 'man/figures' directories.  Otherwise, please put
them in the top-level 'tools' directory, or a subdirectory of it.

The CRAN incoming checks in r-devel were changed to perform the pandoc
conversion checks with '--self-contained', and hence will warn about
missing images.

Pls ensure completeness in the next regular update of your package.”

Fixed using the same figure from the vignette example.