missing comma was preventing parsing of json file
p.s.: Great to see codemeta.json in the wild!  We now have a little R package to assist in generating these files based on your package DESCRIPTION if you want to try it out: https://github.com/codemeta/codemetar

There's a few minor changes to the schema but overall this is very nice!